remove title from logo grid
